requirements.txt
pvporcupine==3.0.2
pyaudio==0.2.14
python-dotenv==1.0.1
requests==2.31.0
openai~=1.74.0